app-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.10.1
httpx==0.27.2
asyncpg==0.28.0
python-dotenv==1.0.1
pyjwt==2.8.0
SQLAlchemy==2.0.21
alembic==1.12.0
